requirements.txt
langchain
llama-index
sentence-transformers
faiss-cpu
openai
transformers
torch
numpy
pandas
openai-whisper
pillow
opencv-python

